requirements.txt
python-binance
TA-Lib
config
numpy
websocket-client
python-dotenv